Scope model container to app
